Move BaseModel to own file
